Refactor buffer/struct mapping and add foreign layouts
Major refactor of buffer mapping and GPU struct layout handling.

* Reworked NativeBuffer/BufferMapping API: removed unmap()/old mapping internals and added address() accessor.
* Buffers now return VirtualBufferMapping instances (new class) for mapped slices instead of exposing internal ByteBuffer/Pointer tracking.
* Introduced VirtualBufferMapping to provide memory-backed views (ByteBuffer/primitive buffers) over MemorySegment or ByteBuffer-backed regions.
* Replaced old StructView/StructLayout usages with a new GpuStruct/GpuStructLayout abstraction and MappedStruct helper. Added ArraySize annotation for fixed-length record component arrays.
* Added foreign memory-based struct support: ForeignStructLayout and ForeignStructLayoutGenerator map Java record-based GpuStructs to java.lang.foreign MemoryLayout (supports STD140/STD430 layout rules and common vector/matrix types).
* Updated StructLayoutGenerator: introduced StructField interface and SimpleStructField, adjusted StructLayout record shape and lookup.
* Removed several legacy buffer-oriented helper fields and pointer buffers (cleaner memory management) and updated destroy actions accordingly.
* Added buffer barrier support: CommandBuffer/CommandDelegate now support buffer memory barriers (both legacy and sync2 variants) and new method overloads; implemented buffer barrier creation in delegates.
* BuiltinDescriptorManager moved to foreign layouts and StructUniform usage, computes initial builtin values and updates per-frame; wiring changes in VulkanRenderer to update descriptor manager and adjust stagingRing upload ordering.
* Miscellaneous small fixes and formatting/cleanup across Vulkan pipeline, command, and staging ring copy logic.

+++ sethlans-core/src/fr/sethlans/core/render/buffer/MemorySize.java
@@ -8,7 +8,7 @@ public final class MemorySize {
     private final int bytesPerElement;
     private final long offset, bytes;
 
-    public static MemorySize bytes(int elements) {
+    public static MemorySize bytes(long elements) {
         return new MemorySize(0, elements, Byte.BYTES);
     }
 
@@ -16,7 +16,7 @@ public final class MemorySize {
         return new MemorySize(offset, elements, Byte.BYTES);
     }
 
-    public static MemorySize shorts(int elements) {
+    public static MemorySize shorts(long elements) {
         return new MemorySize(0, elements, Short.BYTES);
     }
 
@@ -32,7 +32,7 @@ public final class MemorySize {
         return new MemorySize(offset, elements, Integer.BYTES);
     }
 
-    public static MemorySize floats(int elements) {
+    public static MemorySize floats(long elements) {
         return new MemorySize(0, elements, Float.BYTES);
     }
 
@@ -40,7 +40,7 @@ public final class MemorySize {
         return new MemorySize(offset, elements, Float.BYTES);
     }
 
-    public static MemorySize doubles(int elements) {
+    public static MemorySize doubles(long elements) {
         return new MemorySize(0, elements, Double.BYTES);
     }
